renamed form's common.ts to type.ts
