+++ frontend/app/src/components/form/dynamic-form.tsx
@@ -1,8 +1,8 @@
 import { Button } from "@/components/buttons/button-primitive";
 import { Form, FormProps, FormSubmit } from "@/components/ui/form";
+import { DynamicFieldProps } from "@/components/form/type";
 import { SCHEMA_ATTRIBUTE_KIND } from "@/config/constants";
 import CheckboxField from "./fields/checkbox.field";
-import { DynamicFieldProps } from "./fields/common";
 import InputField from "./fields/input.field";
 import PasswordInputField from "./fields/password-input.field";
 import TextareaField from "./fields/textarea.field";
